안지산/final_model1/requirements.txt
torch==1.11.0
torchvision==0.12.0
torchaudio==0.11.0
pytorchvideo==0.1.5
decord==0.6.0
opencv-python
transformers==4.42.0
sentencepiece
tqdm
scikit-learn
pyyaml
soundfile
